Initial commit
This codebase is an extracted internal package taken from
https://github.com/disc/terraform-provider-pritunl, all authored
by Alexandr Hacicheant (github.com/disc), with addition of `Key` struct
as well as `GetUserKeyUrls` and `GetUserKeys` functions which are my
extensions.

I believe this codebase makes a lot of sense as a standalone Go package
rather than an internal package specific to that Terraform provider.